refact: Breaking log module in parts
